requirements.txt
# Hub'Eau Data Integration Pipeline
# Dépendances Python pour Dagster - Version complète

# Core Dagster
dagster==1.8.7
dagster-webserver==1.8.7
dagster-postgres==0.24.8

# Base de données
psycopg[binary]==3.1.13
neo4j==5.15.0

# HTTP et cache
httpx==0.26.0
redis==5.0.1

# S3/MinIO
boto3==1.34.0
botocore==1.34.0

# Data processing
pandas==2.2.0
pyarrow==15.0.0
numpy==1.26.0

# Géospatial
geopandas==0.14.0
shapely==2.0.2
pyproj==3.6.1

# RDF et ontologies
rdflib==7.0.0
owlrl==6.0.2

# WFS et services OGC
owslib==0.29.2
requests==2.31.0

# Utilitaires
python-dotenv==1.0.0
pydantic==2.5.3
python-dateutil==2.8.2

# Monitoring et logs
structlog==23.2.0
prometheus-client==0.19.0

# Mathématiques et statistiques
scipy==1.11.4
scikit-learn==1.3.2

# Tests et développement
pytest==8.0.0
pytest-asyncio==0.23.2
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
